Add setuid wrappers for per-UID buffers
In case of a per-UID buffer, events following a change of user ID should be
forwarded to buffers of the new UID. In order to do so, we add a wrapper around
setuid(3), setreuid(3) and setresuid(2) that unregister and re-register the
application from the session daemon.

While there exist some other functions such as seteuid(3) or setfsuid(2), we
don't need to add a wrapper for them since they do not edit the real user ID,
the actual ID returned by getuid(3) and used for creating the buffers.

Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabriel.pollo-guilbert@efficios.com>
Acked-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>