Stop listener threads while the application is forking
Taken from the fork(2) man page:

	After a fork() in a multithreaded program, the child can safely call
	only async-signal-safe functions (see signal-safety(7)) until such
	time as it calls execve(2).

If an application being traced expects to be single thread and calls fork(2),
the child may be put in a undefined state. Current wrappers around fork(2) in
LTTng-UST takes into account this possibility with various mutexes. Stopping
the listener threads before forking and restarting them in the parent process
simplify these corner cases as no mutex needs to be held during at the time of
the fork since it is now single process.