Join listener threads after sending the cancellation request
Listener threads used to be created in detached state to prevent joining on
a thread blocked on a system call, in our case sys_futex. To prevent that,
we set thread cancellability to asynchronous only for the duration of the
system call. Doing so is safe because UST lock is never held while waiting
on sys_futex.

Most system calls (e.g. read and recv) that needs to be cancellable under
POSIX are implemented in the same way in glibc, by enabling asynchronous
cancellation for the duration of the system call.