WIP: enable asynchronous cancel on futex call
A race condition may occur if we receive a cancellation request between
the futex_async call and the last cancellation point. In this situation, the
thread may not terminate quickly enought since the futex call is blocking.
This case is fixed by temporarily setting the cancel state to asynchronous
during the blocking call.