Use a state machine for UST listener threads
Using a state machine in the listener thread makes them more robust to being
cancelled upon application exit or forking. When a thread is cancelled, it may
only exit outside of the UST lock which is held during critical operation that
shouldn't be interrupted. By keeping the state of what the listener thread is
doing, we can restart the thread later where it was stopped.

This patch also does some cleanup by moving most the critical sections in
separate functions in order to keep the thread main code more readable.